+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased](https://github.com/axa-group/oauth2-mock-server/compare/v1.4.0...HEAD)
+
+### Added
+
+- Add HTTP request object to `OAuth2Service`'s events
+
 ## [1.4.0](https://github.com/axa-group/oauth2-mock-server/compare/v1.3.3...v1.4.0) — 2020-01-15
 
 ### Security